requirements.txt
Pillow
safetensors
torch
